requirements.txt
ansible==10.1.0
ansible-core==2.17.1
bcrypt==4.1.3
cffi==1.16.0
cryptography==42.0.8
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
paramiko==3.4.0
passlib==1.7.4
pycparser==2.22
PyNaCl==1.5.0
PyYAML==6.0.1
resolvelib==1.0.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8

